requirements.txt
Appium-Python-Client==3.2.1
attrs==23.2.0
browserstack-local==1.2.7
certifi==2024.2.2
decorator==5.1.1
docutils==0.20.1
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
kitchen==1.2.6
natsort==8.4.0
outcome==1.3.0.post0
psutil==5.9.8
PySocks==1.7.1
robotframework==7.0
robotframework-appiumlibrary==2.0.0
robotframework-pabot==2.18.0
robotframework-stacktrace==0.4.1
selenium==4.16.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
wsproto==1.2.0
